pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.3.1",
    "google-genai>=1.23.0",
    "google-auth-oauthlib>=1.2.2",
    "google-api-python-client>=2.174.0",
    "sift-stack-py>=0.7.0",
    "yt-dlp>=2025.6.25",
    "moviepy>=2.2.1",
    "requests>=2.32.4",
    "pydantic>=2.11.7",
    "whisper>=1.1.10",
    "sqlalchemy>=2.0.41",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "anthropic>=0.55.0",
    "openai>=1.93.0",
]
